api/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Redis for caching and rate limiting
redis>=5.0.0
hiredis>=2.2.3

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring and logging
prometheus-client>=0.19.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2
locust>=2.17.0
